+++ Flocking Boids/boid.h
@@ -17,14 +17,14 @@ class Boid {
 
 	private:
 
-		std::array<int, 2> velocity;
+		std::array<double, 2> velocity;
 		Point location;
 
 	public:
 
 		Boid();
-		Boid( int, int, int, int );
-		Boid( const Point &, int, int );
+		Boid( int, int, double, double );
+		Boid( const Point &, double, double );
 		// Will get the first two elements of the iterable for location. 
 		template<typename Iter> Boid( int, int, Iter );
 		template<typename Iter> Boid( Iter );
